+++ container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_value_consts/label_value_consts.go
@@ -11,18 +11,21 @@ const (
 	//
 	//   If you add new immutable values to this section, MAKE SURE TO UPDATE THE UNIT TEST!
 	//
-	appIdLabelValueStr               = "kurtosis"
-	engineContainerTypeLabelValueStr = "kurtosis-engine"
+	appIdLabelValueStr                      = "kurtosis"
+	engineContainerTypeLabelValueStr        = "kurtosis-engine"
+	logsDatabaseContainerTypeLabelValueStr  = "kurtosis-logs-db"
+	logsCollectorContainerTypeLabelValueStr = "kurtosis-logs-collector"
 	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! DO NOT CHANGE THESE VALUES !!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-	apiContainerContainerTypeLabelValueStr          = "api-container"
-	userServiceContainerTypeLabelValueStr           = "user-service"
-	networkingSidecarContainerTypeLabelValueStr     = "networking-sidecar"
+	apiContainerContainerTypeLabelValueStr           = "api-container"
+	userServiceContainerTypeLabelValueStr            = "user-service"
+	networkingSidecarContainerTypeLabelValueStr      = "networking-sidecar"
 	moduleContainerTypeLabelValueStr                 = "module"
 	filesArtifactsExpanderContainerTypeLabelValueStr = "files-artifacts-expander"
 
 	enclaveDataVolumeTypeLabelValueStr            = "enclave-data"
 	filesArtifactExpansionVolumeTypeLabelValueStr = "files-artifacts-expansion"
+	logsDatabaseVolumeTypeLabelValueStr           = "logs-db"
 
 	trueValueStr  = "true"
 	falseValueStr = "false"
@@ -36,6 +39,8 @@ const (
 //
 var AppIDDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(appIdLabelValueStr)
 var EngineContainerTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(engineContainerTypeLabelValueStr)
+var LogsDatabaseTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(logsDatabaseContainerTypeLabelValueStr)
+var LogsCollectorTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(logsCollectorContainerTypeLabelValueStr)
 var ModuleContainerTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(moduleContainerTypeLabelValueStr)
 
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! DO NOT CHANGE THESE VALUES !!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -49,3 +54,4 @@ var FilesArtifactExpanderContainerTypeDockerLabelValue = docker_label_value.Must
 
 var EnclaveDataVolumeTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(enclaveDataVolumeTypeLabelValueStr)
 var FilesArtifactExpansionVolumeTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(filesArtifactExpansionVolumeTypeLabelValueStr)
+var LogsDatabaseVolumeTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(logsDatabaseVolumeTypeLabelValueStr)