+++ container-engine-lib/docs/changelog.md
@@ -3,6 +3,9 @@
 * Migrated the `engine functions` from `docker_kurtosis_backend` package to the `engine_functions` package
 * Removed duplicated shared functions that where present in the both the `docker_kurtosis_backend` package and the `shared_helpers` package
 * Added the `consts` package to group the general constants used in the Docker Kurtosis backend
+* Added the `ForLogsDatabaseServer` and the `ForLogsCollectorServer` Docker objects attribute
+* Added `LogsCollectorTypeDockerLabelValue`, `ModuleContainerTypeDockerLabelValue` and `LogsDbDataVolumeTypeDockerLabelValue` Docker labels
+* Added `EngineGUIDDockerLabelKey` Docker label key
 
 # 0.34.0